// api/types.ts
export interface Combatant {
    id: string;
    name: string;
    hp_current: number;
    hp_max: number;
    ac: number;
    initiative: number;
    initiative_bonus: number;
    cr: number;
    is_player: boolean;
}

export interface BattleState {
    combatants: Combatant[];
    turn_order: string[];
    current_turn_index: number;
    round_number: number;
}